requirements.txt
dearpygui
fabric
matplotlib
numpy
opencv-python
pandas
scikit-learn
ffmpeg-python
colour-science
colour-demosaicing
OpenImageIO